lista de lojas e itens na pagina inicial

+++ src/main/java/br/edu/ifsp/lp2a2/comparex/comum/model/entidades/ResultadoQuantidadeLojas.java
@@ -2,9 +2,9 @@ package br.edu.ifsp.lp2a2.comparex.comum.model.entidades;
 
 public class ResultadoQuantidadeLojas {
     private Integer id;
-    private Integer quantidade;
+    private Long quantidade;
 
-    public ResultadoQuantidadeLojas(Integer id, Integer quantidade){
+    public ResultadoQuantidadeLojas(Integer id, Long quantidade){
         this.id = id;
         this.quantidade = quantidade;
     }
@@ -17,11 +17,11 @@ public class ResultadoQuantidadeLojas {
         this.id = id;
     }
 
-    public Integer getQuantidade() {
+    public Long getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(Integer quantidade) {
+    public void setQuantidade(Long quantidade) {
         this.quantidade = quantidade;
     }
 }
